@@ -32,7 +32,7 @@ dependencies:
     sdk: flutter
   cupertino_icons: ^1.0.8
   camera: ^0.10.5
-  tflite_flutter: ^0.10.4
+  tflite_flutter: ^0.11.0
   permission_handler: ^11.0.0
   image: ^4.2.0
   model_viewer_plus: ^1.7.0
@@ -65,6 +65,7 @@ flutter:
   assets:
     - assets/img/
     - assets/glb/
+    - assets/model/
 
 
 
